fix(ticket view): show slack name in comment (#50)
* feat(frontend): improve user display in UserWithAvatar component by showing
"Unknown User" as fallback name, and "Loading..." when user data is not
yet available. Update queries to include user information in comments.
Update types and constants related to tickets and comments for better
clarity and consistency. Add Avatar components for user avatars in
TicketDetailPage. Update schema and resolvers to handle user data in
comments.

* chore(gitignore): update .gitignore to properly exclude frontend/dist contents except .gitkeep